requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-sqlalchemy==3.1.1
pymysql==1.1.1
python-dotenv==1.1.0
gunicorn==23.0.0
